+++ src/_types/YobtaItems.ts
@@ -1,7 +1,5 @@
-import type {
-  PipeFactoryResult,
-  SyncRulesPipeYobta,
-} from '../_internal/pipe/index.js'
+import type { YobtaSyncRule, YobtaSyncRules } from '../ruleYobta/index.js'
+import type { PipeFactoryResult, SyncRulesPipeYobta } from './YobtaPipe.js'
 import type {
   YobtaSyncChain1,
   YobtaSyncChain10,
@@ -15,10 +13,9 @@ import type {
   YobtaSyncChain7,
   YobtaSyncChain8,
   YobtaSyncChain9,
-} from '../_types/YobtaSyncChain.js'
-import type { YobtaSyncRule, YobtaSyncRules } from '../ruleYobta/index.js'
+} from './YobtaSyncChain.js'
 
-export interface ItemsYobta {
+export interface YobtaItems {
   <R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12>(
     ...rules: YobtaSyncChain12<
       R1,