+++ src/createAsyncValidator/createAsyncValidator.ts
@@ -1,6 +1,6 @@
 import { createContext } from '../_internal/createContext/index.js'
 import { handleUnknownError } from '../_internal/parseUnknownError/index.js'
-import type { PipeFactoryResult } from '../_internal/pipe/index.js'
+import type { PipeFactoryResult } from '../_types/YobtaPipe.js'
 import type { YobtaAsyncValidatorResult } from '../_types/YobtaAsyncValidator.js'
 import type { YobtaAsyncValidatorFactory } from '../_types/YobtaAsyncValidatorFactory.js'
 import type { YobtaContext } from '../_types/YobtaContext.js'