+++ src/items/items.ts
@@ -1,19 +1,20 @@
 import { handleUnknownError } from '../_internal/parseUnknownError/index.js'
-import type { Functions, PipeFactoryResult } from '../_internal/pipe/index.js'
-import { pipe } from '../_internal/pipe/index.js'
-import type { ItemsYobta } from '../_types/ItemsYobta.js'
+import type { YobtaItems } from '../_types/YobtaItems.js'
+import type { PipeFactoryResult } from '../_types/YobtaPipe.js'
 import type { YobtaSyncRule, YobtaSyncRules } from '../ruleYobta/index.js'
 import { ruleYobta } from '../ruleYobta/index.js'
 
-export const itemsYobta: ItemsYobta = <R extends YobtaSyncRules>(
+export const items: YobtaItems = <R extends YobtaSyncRules>(
   ...rules: R
 ): YobtaSyncRule<unknown[], PipeFactoryResult<R>[]> => {
   return ruleYobta((input: unknown[], context) => {
-    const next = rules.map(rule => rule(context)) as Functions
-
     return input.map((item, index) => {
       try {
-        return pipe(...next)(item)
+        let next = item
+        for (const rule of rules) {
+          next = rule(context)(next)
+        }
+        return next as PipeFactoryResult<R>
       } catch (error) {
         throw handleUnknownError({
           error,